:art: Move short_location to its own file
