+++ src/rogallo/types.py
@@ -20,7 +20,6 @@ from sybaritic import SpartanURI
 ##############################################################################
 # Wasat imports.
 from wasat import GeminiURI
-from wasat.uri import GEMINI_PREFIX
 
 ##############################################################################
 type RogalloLocation = Path | GeminiURI | FingerURI | GopherURI | SpartanURI
@@ -42,26 +41,6 @@ GEMINI_EXTENSIONS: Final[set[str]] = {DEFAULT_GEMINI_EXTENSION, ".gmni", ".gemin
 ##############################################################################
 class SpartanURINeedingData(SpartanURI):
     """A SpartanURI that requires data to be sent with the request."""
-
-
-##############################################################################
-def short_location(location: RogalloLocation) -> str:
-    """Get a short string representation of a location.
-
-    Args:
-        location: The location to get a short string representation of.
-
-    Returns:
-        A short string representation of the location.
-    """
-    if isinstance(location, (FingerURI, GopherURI, SpartanURI)):
-        return str(location)
-    if isinstance(location, GeminiURI):
-        return str(location).removeprefix(GEMINI_PREFIX)
-    try:
-        return (Path("~") / location.relative_to(Path.home())).as_posix()
-    except ValueError:
-        return location.as_posix()
 
 
 ### types.py ends here
